+++ moviecatalog/src/app/components/filter/filter.component.ts
@@ -1,6 +1,6 @@
-import { distinct, tap, map } from 'rxjs/internal/operators';
-import { Observable, BehaviorSubject, of } from 'rxjs';
-import { MoviesService } from '../services/movies.service';
+import { map } from 'rxjs/internal/operators';
+import { Observable } from 'rxjs';
+import { MoviesService } from './../../services/movies.service';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
