+++ moviecatalog/src/app/components/moviedashboard/moviedashboard.component.ts
@@ -1,5 +1,5 @@
-import { MoviesService } from './../services/movies.service';
-import { Movie } from './../models/movie.interface';
+import { MoviesService } from './../../services/movies.service';
+import { Movie } from './../../core/models/movie.interface';
 import { Component } from '@angular/core';
 import { Observable } from 'rxjs';
