+++ moviecatalog/src/app/core/models/movie.interface.ts
