+++ moviecatalog/src/app/core/models/user.interface.ts
