+++ moviecatalog/src/app/routing/routing.module.ts
@@ -1,11 +1,10 @@
 import { MoviesService } from './../services/movies.service';
-import { MoviesResolve } from './../moviedashboard/moviedashboard.resolve';
-import { AddfilmCloseGuard } from './../addfilm/addfilm.close.guard';
-import { AddFilmGuard } from './../addfilm/addfilm.guard';
-import { PageNotFoundComponent } from './../page-not-found/page-not-found.component';
-import { AddfilmComponent } from './../addfilm/addfilm.component';
-import { AccountComponent } from './../account/account.component';
-import { MoviedashboardComponent } from './../moviedashboard/moviedashboard.component';
+import { AddfilmCloseGuard } from './../components/addfilm/addfilm.close.guard';
+import { AddFilmGuard } from './../components/addfilm/addfilm.guard';
+import { PageNotFoundComponent } from './../components/page-not-found/page-not-found.component';
+import { AddfilmComponent } from './../components/addfilm/addfilm.component';
+import { AccountComponent } from './../components/account/account.component';
+import { MoviedashboardComponent } from './../components/moviedashboard/moviedashboard.component';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes, CanActivate } from '@angular/router';
@@ -14,8 +13,7 @@ const routes: Routes = [
   { path: '', redirectTo: 'main', pathMatch: 'full' },
   {
     path: 'main',
-    component: MoviedashboardComponent,
-    // resolve: { movies: MoviesResolve },
+    component: MoviedashboardComponent
   },
   { path: 'account', component: AccountComponent },
   {
@@ -27,7 +25,9 @@ const routes: Routes = [
   {
     path: 'about-page',
     loadChildren: () =>
-      import('../about-page/about-page.module').then((m) => m.AboutPageModule),
+      import('./../components/about-page/about-page.module').then(
+        (m) => m.AboutPageModule,
+      ),
   },
   { path: '**', component: PageNotFoundComponent },
 ];
@@ -36,6 +36,6 @@ const routes: Routes = [
   declarations: [],
   imports: [CommonModule, RouterModule.forRoot(routes)],
   exports: [RouterModule],
-  providers: [AddFilmGuard, AddfilmCloseGuard, MoviesService, MoviesResolve],
+  providers: [AddFilmGuard, AddfilmCloseGuard, MoviesService],
 })
 export class RoutingModule {}
